+++ release.sh
@@ -1,0 +1,18 @@
+#! /usr/bin/env bash
+
+set -eu -o pipefail
+
+# Verify that PyPi package display is valid
+echo "Checking PyPi package HTML"
+if ! python setup.py --long-description | rst2html.py --strict > /dev/null; then
+    echo
+    echo "Error: invalid package display; check CHANGELOG.md"
+    exit 1
+fi
+
+# Build package
+rm -rf build dist
+python setup.py sdist bdist_wheel
+
+# Upload to PyPi
+twine upload -u rax_cbd_dev dist/*